Booking service wip

+++ src/main/java/com/acker/busticketbackend/buses/UpdateFareRequest.java
@@ -12,5 +12,5 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 @NoArgsConstructor
 public class UpdateFareRequest {
-    private float pricePerUnitDistance;
+    private Double pricePerUnitDistance;
 }
